+++ packages/azure-themes/src/azure/IAzureSemanticColors.ts
@@ -63,4 +63,8 @@ export interface IAzureSemanticColors {
     hover: string;
     select: string;
   };
+  shimmer: {
+    primary: string;
+    secondary: string;
+  };
 }