+++ packages/dashboard/src/components/Card/Card.tsx
@@ -34,7 +34,8 @@ export class Card extends React.Component<ICardProps, ICardState> {
       actionBarOverflowButtonTitle,
       actionBarOverflowButtonAriaLabel,
       actionBarOverflowButtonAriaDescription,
-      role
+      role,
+      cardAriaLabelledby
     } = this.props;
     const getClassNames = classNamesFunction<ICardProps, ICardStyles>();
     const classNames = getClassNames(getStyles);
@@ -42,7 +43,7 @@ export class Card extends React.Component<ICardProps, ICardState> {
     // in DGL where it explicitly had direction set to LTR, which was overriding card's RTL behavior
     const direction = getRTL() ? 'rtl' : 'ltr';
     return (
-      <div dir={direction} className={classNames.root} role={role ? role : 'region'}>
+      <div dir={direction} className={classNames.root} role={role ? role : 'region'} aria-labelledby={cardAriaLabelledby}>
         <CardFrame
           cardTitle={cardFrameContent.cardTitle}
           cardDropDownOptions={cardFrameContent.cardDropDownOptions}