+++ packages/experiments/src/components/Persona/Vertical/VerticalPersona.types.ts
@@ -1,5 +1,5 @@
 import { IStyle, IFontWeight } from '@uifabric/styling';
-import { IComponent, IHTMLDivSlot, IStyleableComponentProps } from '../../../Foundation';
+import { IComponent, IHTMLSlot, IStyleableComponentProps } from '../../../Foundation';
 import { ITextSlot } from '../../../Text';
 import { IPersonaCoinSlot } from '../../PersonaCoin/PersonaCoin.types';
 
@@ -12,7 +12,7 @@ export type IVerticalPersonaTokenReturnType = ReturnType<Extract<IVerticalPerson
 export type IVerticalPersonaStylesReturnType = ReturnType<Extract<IVerticalPersonaComponent['styles'], Function>>;
 
 export interface IVerticalPersonaSlots {
-  root?: IHTMLDivSlot;
+  root?: IHTMLSlot;
   primaryText?: ITextSlot;
   secondaryText?: ITextSlot;
   coin?: IPersonaCoinSlot;