+++ packages/experiments/src/components/Stack/examples/Stack.Horizontal.WrapNested.Example.tsx
@@ -75,8 +75,8 @@ export class HorizontalStackWrapNestedExample extends React.Component<{}, IExamp
           onChange={this._onWidthChange}
         />
 
-        <Stack horizontal wrap gap={40} verticalGap={30} className={styles.root}>
-          <Stack horizontal wrap gap={30} verticalGap={10} className={styles.stackOne}>
+        <Stack horizontal wrap gap={'30 40'} className={styles.root}>
+          <Stack horizontal wrap gap={'10 30'} className={styles.stackOne}>
             <span>1</span>
             <span>2</span>
             <span>3</span>
@@ -86,7 +86,7 @@ export class HorizontalStackWrapNestedExample extends React.Component<{}, IExamp
             <span>7</span>
           </Stack>
 
-          <Stack horizontal wrap gap={50} verticalGap={20} className={styles.stackTwo}>
+          <Stack horizontal wrap gap={'20 50'} className={styles.stackTwo}>
             <span>1</span>
             <span>2</span>
             <span>3</span>