+++ packages/experiments/src/slots/examples/Slots.Root.Example.tsx
@@ -8,7 +8,7 @@ export class SlotsRootExample extends React.Component<{}, {}> {
     return (
       <Stack {...stackProps}>
         <Button icon="share" href="https://developer.microsoft.com/en-us/fabric" content="Root: Implicit 'a' via href prop" />
-        <Button icon="share" root={(rootProps, RootType) => <RootType {...rootProps} />} content="Root: Function" />
+        <Button icon="share" root={render => render((RootType, rootProps) => <RootType {...rootProps} />)} content="Root: Function" />
       </Stack>
     );
   }