+++ packages/office-ui-fabric-react/src/components/ColorPicker/ColorPicker.types.ts
@@ -2,7 +2,10 @@ import { ITheme, IStyle } from '../../Styling';
 import { IBaseProps, IRefObject, IStyleFunctionOrObject } from '../../Utilities';
 import { IColor } from '../../utilities/color/colors';
 
-export interface IColorPicker {}
+export interface IColorPicker {
+  /** The currently selected color. */
+  color: IColor;
+}
 
 export interface IColorPickerProps extends IBaseProps<IColorPicker> {
   /**
@@ -16,35 +19,35 @@ export interface IColorPickerProps extends IBaseProps<IColorPicker> {
   color: string;
 
   /**
-   * Callback issued when the user changes the color.
+   * Callback for when the user changes the color.
    */
   onColorChanged?: (color: string, colorObject: IColor) => void;
 
   /**
-   * The setting of whether to hide the alpha control slider.
+   * Whether to hide the alpha control slider.
    */
   alphaSliderHidden?: boolean;
 
   /**
-   * Label for the hex textfield.
+   * Label for the hex text field.
    * @defaultvalue Hex
    */
   hexLabel?: string;
 
   /**
-   * Label for the red textfield.
+   * Label for the red text field.
    * @defaultvalue Red
    */
   redLabel?: string;
 
   /**
-   * Label for the green textfield.
+   * Label for the green text field.
    * @defaultvalue Green
    */
   greenLabel?: string;
 
   /**
-   * Label for the blue textfield.
+   * Label for the blue text field.
    * @defaultvalue Blue
    */
   blueLabel?: string;